commented out code which causes a failure
I commended out the code I put into the IndexUnitTest
method, to cause a divide by zero error. I put that there
a week or so ago, just to be sure that an error would
occur. Becausee, without it this xUnit test passed in
Azure DevOps, which doesn't make sense to me
since it doesn't have access to the backend
database server which is on-prem.